+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Add optional useRawFile parameter into PersistenceSettings.Load functions
+
 ### Removed
 - IPersistenceSettings interface
+- Static PersistenceSettings.LoadRawFile function
 
 ## [2.2.0] - 2023-03-28
 ### Added